migrate tags from charfiled to new pg array field.
We added 4 migrations. The first one adds a new field tags_arr which is a pg array field.
The seconds one is a data migration which takes all tags from the previous
tags charfield, parses it and adds it to the new tags_arr field.
The Third one removes the old tags field. And the last, fourth one,
renames tags_arr to tags.