more members app refactorings
using some of the django admin UserAdmin mechanisms for handling
member user creation and change. as with contrib.user we know only
create a user with username and password combo. Then we are redirected to the
change view where we fill out the rest of the info. The password is not displayed
in the change view but we utilize a password change form from contrib.user.
We use md5 for hashing password and secret_key for salting. it's not perfect
but it's better than before.

Also, we removed the contact_type table as there is not need for it.
Insted we added 3 Types to the Contact model: email, phone, cell.
A data migration that migrates old data from the contact_type table
to the new charfile is also in place.