got sick of subpages app
there's no actual need for this to be in the db cause nobody edits it for years
and it makes it harder to setup a local copy for developemnt, not to mention versioning
a bunch of media/uploads files. Subpages is no completely static, with nothing saved in the db.