Adds new Makefile targets
These are for convenience for creating a data dir on the host
for the db container.

Use 'make db' to create a data dir "__data" for storing the postgres
data from the db container. The fig file will use this directory.

Use "make db-restore" to import data from a databse dump. The database
dump file must be named "ksetdb.sql".

There's also "make db-prompt" if you wish to get a psql shell in the
contianer.

To start the project run "docker-compose up" (or fig up if you're
running an older version of fig)